reader: update RepoConfig
As it turns out, I don't need the `new` method
for the implementation of `RepoConfig` struct.

What I need really is just the default which is done
by creating a `Default` type and `default` method for the
`Default` type for `RepoConfig` which does the job
of setting the defaults.

The next goal would be to create a `write_config` method
for default openSUSE repositories but the question is, is that
even needed? I have a feeling that this very compatible
with RPM-based distributions so I have to hold this out for now.
We can always check `/etc/os-release` or
`/etc/lsb_release` (deprecated) to write the config but I don't
think this will happen and will cause more burden than
what is necessary.